+++ CHANGELOG.md
@@ -6,5 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v1.1.1] - 2020-12-14
+
+### Added
+- Add methods to install, purge, upgrade and start dogus
+- Add methods to run maven and yarn integration tests
+- Add parameters to test exection with yarn and maven
+  - boolean flag to control video recording
+  - argument list to pass additional environment variables
+
 ## [v1.0.0] - 2020-06-09
 First release of the dogu-build-lib!